Allow requests to be gracefully canceled
The "oneshot" sender has had support for this for a while, but it wasn't
plumbed through.

Prior to this commit, if the tokio runtime dropped a "render in BG
thread" request, it would still be rendered- blocking rendering of a
reloaded page, if one was present.

This makes it so that _if_ the runtime drops the request processing,
we'll cancel some of the other processing relatively promptly (at the
end of the current row.)

(It's not clear to me how to convince Axum/Tokio/Hyper to drop the
request...but this does seem more responsive.)